requirements.txt
alembic==0.9.7
bcrypt==3.1.4
blinker==1.4
cffi==1.11.4
click==6.7
dominate==2.3.1
Faker==0.8.10
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.18
PyMySQL==0.8.0
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.1
text-unidecode==1.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
